+++ zcash_client_sqlite/src/wallet/orchard.rs
@@ -894,6 +894,17 @@ pub(crate) mod tests {
     }
 
     #[test]
+    fn tip_shard_notes_remain_spendable_after_chain_tip_update() {
+        testing::pool::tip_shard_notes_remain_spendable_after_chain_tip_update::<OrchardPoolTester>(
+        )
+    }
+
+    #[test]
+    fn unscanned_range_below_anchor_blocks_spendability() {
+        testing::pool::unscanned_range_below_anchor_blocks_spendability::<OrchardPoolTester>()
+    }
+
+    #[test]
     fn ovk_policy_prevents_recovery_from_chain() {
         testing::pool::ovk_policy_prevents_recovery_from_chain::<OrchardPoolTester>()
     }